+++ pv_evaluation/estimators/__init__.py
@@ -6,12 +6,12 @@ from pv_evaluation.estimators.cluster_precision import cluster_precision_estimat
 from pv_evaluation.estimators.cluster_recall import cluster_recall_estimator, cluster_recall_std
 
 __all__ = [
-    "pairwise_precision_estimator", 
-    "pairwise_precision_std", 
-    "pairwise_recall_estimator", 
+    "pairwise_precision_estimator",
+    "pairwise_precision_std",
+    "pairwise_recall_estimator",
     "pairwise_recall_std",
     "cluster_precision_estimator",
     "cluster_precision_std",
     "cluster_recall_estimator",
     "cluster_recall_std",
-    ]
+]