+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added cluster precision and cluster recall estimators (for cluster block sampling).
 - Added additional inventor benchmark datasets.
 
 ## 0.0.1 (June 27, 2022)