+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- BREAKING: remove Harvard inventors benchmark (low data quality).
 - BREAKING: change "mention-id" to "mention_id" and "unique-id" to "unique_id" in datasets.
 - Added inventor names to raw csv benchmarkd datasets (useful for validation and debugging).
 - Added cluster precision and cluster recall estimators (for cluster block sampling).