+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-- Add script for processing hand-disambiguated inventor benchmark.
-- BREAKING: remove Harvard inventors benchmark (low data quality).
+- Added estimates table and plots.
+- FIX: Install from Github now includes necessary csv files.
+- Added script for processing hand-disambiguated inventor benchmark.
+- Removed Harvard inventors benchmark (low data quality).
 - BREAKING: change "mention-id" to "mention_id" and "unique-id" to "unique_id" in datasets.
 - Added inventor names to raw csv benchmarkd datasets (useful for validation and debugging).
 - Added cluster precision and cluster recall estimators (for cluster block sampling).